- Modified App.vue
I have added a link to the new /ibrowse URL - note this uses <router-link> tag which means that the Vue-router will handle the URL when it is loaded. The URLs that Vue Router handles are defined in hello-world/src/router/index.js, as you can see there is also a modification in that file:

- Modified hello-world/src/router/index.js

I have added a new route in here to /ibrowse - this will load up the new component in views/iBrowse.vue

I have also added a route to load an individual site (/ibrowse/sites/:site). This will load up the same views/iBrowse.vue component but will also set the 'site' property.

- Added iBrowse.vue

This component handles the iBrowse "tab" when it is clicked. Please look at the items within the <script> tags. They are important. Interesting ones to look at are:

mounted: these methods will be called when the component is mounted for the first time.
watch: these methods will be called whenever the $route variable changes. This handles resetting the component when it is already mounted,and is reused for another route.

Methods:
loadSiteList: This is used to load a list of sites. It could be an API call
setSiteObject: This is used to find the correct site object when the component has been loaded with the 'site' property set (e.g. when the URL /ibrowse/sites/:site is called)

You will also see here that we include the SiteList component (SiteList.vue) and we call it from within the <template> tags and also pass in the 2 properties it expects.

- Added SiteList.vue

This is a simple component which takes a list of Sites as a property (and also optionally a 'selected' site) and will list them. For the purposes of this example it simply takes the siteList and calls another Component, SiteRow to list out each row of a table.

It will then also render the SiteOverview component (but only if the site property has been passed)

- Added SiteRow.vue
- Added SiteOverview.vue

These components simply take a property and print out their information within the template. SiteRow is interesting because it also generates some <router-link> links in order to show the SiteOverview.